refactor(rsc): drop 'use client' where unneeded and split auth shells to RSC
Remove 'use client' from button.tsx, payment/canceled/content.tsx, and
link-indicator.tsx where no client APIs are used. Convert premium/badge
and profile-analytics to async server components using getTranslation.
Split privacy page and auth/signin layout into server shells that render a
small <PageTrack> client island (and TrackedGithubLink for signin).
