+++ src/app/payment/canceled/content.tsx
@@ -1,4 +1,3 @@
-'use client'
 import Link from 'next/link'
 
 function CanceledPageContent() {
