+++ src/components/button/button.tsx
@@ -1,5 +1,3 @@
-'use client'
-
 import { Loader2 } from 'lucide-react'
 import type React from 'react'
